メンバー編集機能を配列タイプにも対応
